feat(app navigation, font, components):
Added custom font, 'Ubuntu-Regular', will go back and choose an official 
font. This was just for learning purposes.

Added a custom buttom to be used in the Reports screen. Also modified 
the header to receive a title to make it a default header for other 
screens.

Added stack navigation for learning purposes, and then added a tab 
navigation for the main part of the app.